--- src/Nacho.java
@@ -1,10 +1,0 @@
-public class Nacho extends Snack{
-    public Nacho() {
-        this.nombre = "Nacho";
-    }
-
-    @Override
-    public double getPrecio() {
-        return 3.95;
-    }
-}
