+++ src/Products/Botella.java
@@ -1,11 +1,16 @@
-public class Vaso extends Bebida {
+package Products;
 
-    public Vaso(String nombre, double precio, String tamanio, String sabor) {
+import Products.Bebida;
+
+public class Botella extends Bebida {
+
+    public Botella(String nombre, double precio, String tamanio, String sabor) {
         this.nombre = nombre;
         this.precio = precio;
         this.tamanio = tamanio;
         this.sabor = sabor;
     }
+
 
     @Override
     public double getPrecio() {
@@ -21,6 +26,5 @@ public class Vaso extends Bebida {
                 "Sabor: " + sabor + "\n";
 
         return salida;
-
     }
 }
