+++ src/Products/HotDog.java
@@ -1,6 +1,10 @@
+package Products;
+
+import Products.Snack;
+
 public class HotDog extends Snack {
     public HotDog() {
-        this.nombre = "HotDog";
+        this.nombre = "Products.HotDog";
     }
 
     @Override
