+++ src/Products/Nacho.java
@@ -1,0 +1,14 @@
+package Products;
+
+import Products.Snack;
+
+public class Nacho extends Snack {
+    public Nacho() {
+        this.nombre = "Products.Nacho";
+    }
+
+    @Override
+    public double getPrecio() {
+        return 3.95;
+    }
+}
