+++ src/Products/Snack.java
@@ -1,0 +1,5 @@
+package Products;
+
+public abstract class Snack extends Producto {
+
+}
