--- src/Snack.java
@@ -1,3 +1,0 @@
-public abstract class Snack extends Producto{
-
-}
